apps/backend/pyproject.toml
[tool.poetry]
name = "codegraph-backend"
version = "0.1.0"
description = "AI coding agent platform backend"
authors = ["CodeGraph Contributors"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.126.0"
uvicorn = {extras = ["standard"], version = "^0.39.0"}
sqlalchemy = "^2.0.23"
asyncpg = "^0.31.0"
alembic = "^1.12.1"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.21"
langchain = "^1.2.0"
langchain-anthropic = "^1.3.0"
langgraph = "^1.0.5"
langsmith = "^0.5.0"
redis = "^7.1.0"
httpx = "^0.28.1"
structlog = "^25.5.0"
email-validator = "^2.3.0"
bcrypt = "^4.2.0"
aiosmtplib = "^5.0.0"
jinja2 = "^3.1.6"
pyotp = "^2.9.0"
qrcode = {extras = ["pil"], version = "^8.2"}
authlib = "^1.6.6"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"
black = "^25.12.0"
ruff = "^0.14.10"
mypy = "^1.7.1"
types-python-jose = "^3.5.0.20250531"
types-passlib = "^1.7.7.20250602"
aiosqlite = "^0.22.0"
testcontainers = "^4.13.3"
psycopg2-binary = "^2.9.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.12"
strict = true
pythonpath = ["."]
namespace_packages = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"
